+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/helpers/MockDeploymentManager.scala
@@ -66,10 +66,14 @@ class MockDeploymentManager(
   // Pass correct deploymentId
   private def fallbackDeploymentId = DeploymentId(UUID.randomUUID().toString)
 
-  override def getFreshProcessStates(name: ProcessName): Future[List[StatusDetails]] = {
+  override def getProcessStates(
+      name: ProcessName
+  )(implicit freshnessPolicy: DataFreshnessPolicy): Future[WithDataFreshnessStatus[List[StatusDetails]]] = {
     Future {
       Thread.sleep(delayBeforeStateReturn.toMillis)
-      managerProcessStates.getOrDefault(name, prepareProcessState(defaultProcessStateStatus, fallbackDeploymentId))
+      WithDataFreshnessStatus.fresh(
+        managerProcessStates.getOrDefault(name, prepareProcessState(defaultProcessStateStatus, fallbackDeploymentId))
+      )
     }
   }
 
@@ -258,7 +262,8 @@ class MockManagerProvider(deploymentManager: DeploymentManager = new MockDeploym
   override def createDeploymentManager(
       modelData: BaseModelData,
       deploymentManagerDependencies: DeploymentManagerDependencies,
-      deploymentConfig: Config
+      deploymentConfig: Config,
+      scenarioStateCacheTTL: Option[FiniteDuration]
   ): ValidatedNel[String, DeploymentManager] =
     valid(deploymentManager)